+++ src/container/app/comp-app.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function App() {
+  return <h1>App</h1>;
+}
+
+export default App;
