+++ src/container/app/portfoliomod/commons/visualize.js
@@ -1,7 +1,7 @@
 import { fetchCSVData, fetchJSONData } from "./fetch-data";
 import { printError } from "./print-error";
 
-import { visualizePokemon, pokemonFetchURL } from "./../components/pokemon-linear-viz";
+import { visualizePokemon, pokemonFetchURL } from "../components/pokemon-linear-viz";
 
 export function vizPokemon(vizAttr) {
   fetchCSVData(pokemonFetchURL()).then(
