+++ src/container/comp-container.js
@@ -1,7 +1,8 @@
 import React from 'react';
+import App from './app/comp-app'
 
 function Container() {
-  return <div>Hello</div>;
+  return <App />
 }
 
 export default Container;
